Service

+++ src/Model/Produkten.java
@@ -3,10 +3,10 @@ package Model;
 public class Produkten {
 
     private String name;
-    private int price;
+    private double price;
     private String region;
 
-    public Produkten(String name, int price, String region) {
+    public Produkten(String name, double price, String region) {
         this.name = name;
         this.price = price;
         this.region = region;
@@ -20,11 +20,11 @@ public class Produkten {
         this.name = name;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(double price) {
         this.price = price;
     }
